+++ xLights/sequencer/tabSequencer.cpp
@@ -1166,7 +1166,7 @@ void xLightsFrame::EffectFileDroppedOnGrid(wxCommandEvent& event)
     std::string filename = parms[1].ToStdString();
 
     int effectIndex = 0;
-    for (int i = 0; i < EffectsPanel1->EffectChoicebook->GetChoiceCtrl()->GetCount(); i++)
+    for (size_t i = 0; i < EffectsPanel1->EffectChoicebook->GetChoiceCtrl()->GetCount(); i++)
     {
         if (EffectsPanel1->EffectChoicebook->GetChoiceCtrl()->GetString(i) == effectName)
         {
